# Provide "decode" code generation for the streaming variants for all other types