… gen/internal/tests/collision/collision.go
… gen/internal/tests/containers/containers.go
… gen/internal/tests/exceptions/exceptions.go
… gen/internal/tests/hyphenated_file/hyphenated_file.go
… gen/internal/tests/non_hyphenated/non_hyphenated.go
… gen/internal/tests/nozap/nozap.go
… gen/internal/tests/services/services.go
… gen/internal/tests/set_to_slice/set_to_slice.go
… gen/internal/tests/structs/structs.go
… gen/internal/tests/typedefs/typedefs.go
… gen/internal/tests/unions/unions.go
… gen/internal/tests/uuid_conflict/uuid_conflict.go
… gen/roundtrip_test.go
… gen/service_test.go
… gen/struct_test.go
… gen/typedef_test.go
… gen/util_for_test.go
… plugin/api/api.go