#495 provided the code generation for streaming reads of enums off of the wire.
Provide the streaming read generators for all of the other types.

Doing so meant providing a `StreamGenerator` that plumbs into the existing
generators.  This `StreamGenerator` only provides the "decoding" mechanism,
leveraging #491 for reading the raw Thrift encoding off of the wire.

In addition to providing a `StreamGenerator`, the list, map, set, enum, typdef,
and struct generators all added a `Decoder` method that will appropriately
recurse and iterate to generate the proper, mirrored readers for the raw wire
representation.

The `Decode` and `DecodePtr` methods for `StreamGenerator` then hook into the
templated generator itself, providing the templated `decode` and `decodePtr`
calls that will be used where necessary.

The existing tests were leveraged to make sure that the streaming reads were
compatible with the current binary writes and that no data was lost.  In doing
so, a 'genericized' function that will perform the cross-products of
encoding/writing and decoding/reading.

A benchmark was also added to evaluate the new streaming reads.  The following
represents benchmarks of the original code vs the dedup'd code (the binary
decode/encode leveraging the streaming implementations):
```
name                                        old time/op    new time/op    delta
RoundTrip/PrimitiveOptionalStruct/Encode-8    1.76µs ± 2%    1.86µs ± 5%    +5.93%  (p=0.008 n=5+5)
RoundTrip/PrimitiveOptionalStruct/Decode-8    2.55µs ± 3%    2.82µs ± 5%   +10.68%  (p=0.008 n=5+5)
RoundTrip/Graph/Encode-8                      3.15µs ± 2%    3.16µs ± 1%      ~     (p=1.000 n=5+5)
RoundTrip/Graph/Decode-8                      5.02µs ± 1%    8.95µs ±11%   +78.05%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Encode-8     19.1µs ± 1%    19.8µs ± 3%    +3.76%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Decode-8     45.2µs ± 3%   110.8µs ± 1%  +144.98%  (p=0.008 n=5+5)

name                                        old alloc/op   new alloc/op   delta
RoundTrip/PrimitiveOptionalStruct/Encode-8      704B ± 0%      704B ± 0%      ~     (all equal)
RoundTrip/PrimitiveOptionalStruct/Decode-8    1.40kB ± 0%    1.45kB ± 0%    +3.43%  (p=0.008 n=5+5)
RoundTrip/Graph/Encode-8                      1.70kB ± 0%    1.70kB ± 0%      ~     (all equal)
RoundTrip/Graph/Decode-8                      2.78kB ± 0%    3.55kB ± 0%   +27.98%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Encode-8     1.30kB ± 0%    1.30kB ± 0%      ~     (all equal)
RoundTrip/ContainersOfContainers/Decode-8     12.3kB ± 0%    28.6kB ± 0%  +132.07%  (p=0.008 n=5+5)

name                                        old allocs/op  new allocs/op  delta
RoundTrip/PrimitiveOptionalStruct/Encode-8      1.00 ± 0%      1.00 ± 0%      ~     (all equal)
RoundTrip/PrimitiveOptionalStruct/Decode-8      14.0 ± 0%      15.0 ± 0%    +7.14%  (p=0.008 n=5+5)
RoundTrip/Graph/Encode-8                        11.0 ± 0%      11.0 ± 0%      ~     (all equal)
RoundTrip/Graph/Decode-8                        32.0 ± 0%      63.0 ± 0%   +96.88%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Encode-8       18.0 ± 0%      18.0 ± 0%      ~     (all equal)
RoundTrip/ContainersOfContainers/Decode-8        164 ± 0%       837 ± 0%  +410.37%  (p=0.008 n=5+5)
```
and the original code vs the streaming implementations (benchmarks renamed for
the purpose of benchstat-ing).
```
name                                        old time/op    new time/op    delta
RoundTrip/PrimitiveOptionalStruct/Encode-8    1.76µs ± 2%    1.61µs ± 2%   -8.57%  (p=0.008 n=5+5)
RoundTrip/PrimitiveOptionalStruct/Decode-8    2.55µs ± 3%    1.19µs ± 0%  -53.36%  (p=0.008 n=5+5)
RoundTrip/Graph/Encode-8                      3.15µs ± 2%    1.88µs ± 1%  -40.29%  (p=0.008 n=5+5)
RoundTrip/Graph/Decode-8                      5.02µs ± 1%    1.87µs ± 2%  -62.76%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Encode-8     19.1µs ± 1%    12.5µs ± 4%  -34.30%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Decode-8     45.2µs ± 3%    27.6µs ± 1%  -38.90%  (p=0.008 n=5+5)

name                                        old alloc/op   new alloc/op   delta
RoundTrip/PrimitiveOptionalStruct/Encode-8      704B ± 0%       24B ± 0%  -96.59%  (p=0.008 n=5+5)
RoundTrip/PrimitiveOptionalStruct/Decode-8    1.40kB ± 0%    0.10kB ± 0%  -92.57%  (p=0.008 n=5+5)
RoundTrip/Graph/Encode-8                      1.70kB ± 0%    0.02kB ± 0%  -98.59%  (p=0.008 n=5+5)
RoundTrip/Graph/Decode-8                      2.78kB ± 0%    0.22kB ± 0%  -92.22%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Encode-8     1.30kB ± 0%    0.02kB ± 0%  -98.16%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Decode-8     12.3kB ± 0%    10.2kB ± 0%  -17.53%  (p=0.016 n=5+4)

name                                        old allocs/op  new allocs/op  delta
RoundTrip/PrimitiveOptionalStruct/Encode-8      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
RoundTrip/PrimitiveOptionalStruct/Decode-8      14.0 ± 0%      11.0 ± 0%  -21.43%  (p=0.008 n=5+5)
RoundTrip/Graph/Encode-8                        11.0 ± 0%       1.0 ± 0%  -90.91%  (p=0.008 n=5+5)
RoundTrip/Graph/Decode-8                        32.0 ± 0%      11.0 ± 0%  -65.62%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Encode-8       18.0 ± 0%       1.0 ± 0%  -94.44%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Decode-8        164 ± 0%       146 ± 0%  -10.73%  (p=0.008 n=5+5)
```