@witriew 2021-06-16
A couple of things to the reviewers:
* The cross-product testing probably needs a bit of modification to take in additional tests if needed (i.e., testing for equality to the original, which needs to be done after each combination).
* The set/map/list generators need a check added to make sure that the type that's read is the type expected.