@abhinav 2021-06-22
Also, check out [gen/quick_test](https://github.com/thriftrw/thriftrw-go/blob/dev/gen/quick_test.go) at some point. If we add another test to that test suite, it'll validate all generated types in the ThriftRW codebase.